+++ residualConnectivityCommon/subObs/articulationConditioningSameCountImportance.cpp
@@ -64,7 +64,7 @@ namespace residualConnectivity
 					if(nComponentsWithFixedOnVertices > 1)
 					{
 						state[*i].state = FIXED_ON;
-						conditioningVertices.push_back(*i);
+						conditioningVertices.push_back((int)*i);
 					}
 				}
 			}
@@ -114,7 +114,7 @@ namespace residualConnectivity
 			}
 			samplingArgsOriginal.n = nUpVertices;
 			int nDeterministic = 0, nZeroWeights = 0;
-			sampling::samplingBase(samplingArgsOriginal.n, samplingArgsOriginal.indices, samplingArgsOriginal.weights, samplingArgsOriginal.zeroWeights, samplingArgsOriginal.deterministicInclusion, nDeterministic, nZeroWeights);
+			sampling::samplingBase((int)samplingArgsOriginal.n, samplingArgsOriginal.indices, samplingArgsOriginal.weights, samplingArgsOriginal.zeroWeights, samplingArgsOriginal.deterministicInclusion, nDeterministic, nZeroWeights);
 			sampling::computeExponentialParameters(samplingArgsOriginal);
 			calculateExpNormalisingConstants(samplingArgsOriginal);
 
@@ -143,7 +143,7 @@ namespace residualConnectivity
 			}
 			samplingArgsImportance.n = nUpVertices;
 			nDeterministic = 0, nZeroWeights = 0;
-			sampling::samplingBase(samplingArgsImportance.n, samplingArgsImportance.indices, samplingArgsImportance.weights, samplingArgsImportance.zeroWeights, samplingArgsImportance.deterministicInclusion, nDeterministic, nZeroWeights);
+			sampling::samplingBase((int)samplingArgsImportance.n, samplingArgsImportance.indices, samplingArgsImportance.weights, samplingArgsImportance.zeroWeights, samplingArgsImportance.deterministicInclusion, nDeterministic, nZeroWeights);
 			sampling::computeExponentialParameters(samplingArgsImportance);
 			calculateExpNormalisingConstants(samplingArgsImportance);
 
@@ -360,14 +360,14 @@ namespace residualConnectivity
 			int nDeterministic = 0, nZeroWeights = 0;
 			const std::vector<mpfr_class> operationalProbabilities = contextObj.getOperationalProbabilities();
 			for(std::vector<int>::iterator i = newSimulationOrder.begin(); i != newSimulationOrder.end(); i++) newSamplingArgsOriginal.weights.push_back(operationalProbabilities[*i]);
-			sampling::samplingBase(newSamplingArgsOriginal.n, newSamplingArgsOriginal.indices, newSamplingArgsOriginal.weights, newSamplingArgsOriginal.zeroWeights, newSamplingArgsOriginal.deterministicInclusion, nDeterministic, nZeroWeights);
+			sampling::samplingBase((int)newSamplingArgsOriginal.n, newSamplingArgsOriginal.indices, newSamplingArgsOriginal.weights, newSamplingArgsOriginal.zeroWeights, newSamplingArgsOriginal.deterministicInclusion, nDeterministic, nZeroWeights);
 			sampling::computeExponentialParameters(newSamplingArgsOriginal);
 			calculateExpNormalisingConstants(newSamplingArgsOriginal);
 
 			for(std::vector<int>::iterator i = newSimulationOrder.begin(); i != newSimulationOrder.end(); i++) newSamplingArgsImportance.weights.push_back((*importanceProbabilities)[*i]);
 			nDeterministic = 0;
 			nZeroWeights = 0;
-			sampling::samplingBase(newSamplingArgsImportance.n, newSamplingArgsImportance.indices, newSamplingArgsImportance.weights, newSamplingArgsImportance.zeroWeights, newSamplingArgsImportance.deterministicInclusion, nDeterministic, nZeroWeights);
+			sampling::samplingBase((int)newSamplingArgsImportance.n, newSamplingArgsImportance.indices, newSamplingArgsImportance.weights, newSamplingArgsImportance.zeroWeights, newSamplingArgsImportance.deterministicInclusion, nDeterministic, nZeroWeights);
 			sampling::computeExponentialParameters(newSamplingArgsImportance);
 			calculateExpNormalisingConstants(newSamplingArgsImportance);
 			
@@ -388,7 +388,7 @@ namespace residualConnectivity
 				//Number of vertices that are being simulated in this biconnected component
 				std::vector<context::inputGraph::vertex_descriptor>& unFixedVerticesThisComponent = unFixedPointsPerComponent[i];
 				std::vector<context::inputGraph::vertex_descriptor>& fixedVerticesThisComponent = fixedPointsPerComponent[i];
-				int currentComponentSize = unFixedVerticesThisComponent.size();
+				int currentComponentSize = (int)unFixedVerticesThisComponent.size();
 				int copiedSkipped = skipped;
 				//Loop over all the particles
 				for(int j = 0; j < nSimulations; j++)
@@ -443,7 +443,7 @@ namespace residualConnectivity
 				//Update estimate by the fraction of particles for which the current component was connected. 
 				estimate *= (double)sumWeights / (double)nSimulations;
 				//Stratified sampling to bring the number of samples up to nSimulations
-				int multiple = nSimulations / otherNChosen.size();
+				int multiple = nSimulations / (int)otherNChosen.size();
 				nChosen.clear();
 				for(int j = 0; j < multiple; j++)
 				{
